Move util/ to promises/
