+++ promises/test.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
 
 import { assert, test } from "../testing/mod.ts";
-import { defer, isDeferred } from "./deferred.ts";
+import { defer, isDeferred } from "./mod.ts";
 
 test(async function asyncIsDeferred() {
   const d = defer();
