+++ test.ts
@@ -2,18 +2,17 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
 
 import "benching/test.ts";
-import "util/deferred_test.ts";
 import "colors/test.ts";
 import "datetime/test.ts";
 import "examples/test.ts";
 import "flags/test.ts";
+import "fs/path/test.ts";
+import "fs/walk_test.ts";
 import "io/bufio_test.ts";
 import "io/ioutil_test.ts";
 import "io/util_test.ts";
 import "io/writers_test.ts";
 import "io/readers_test.ts";
-import "fs/path/test.ts";
-import "fs/walk_test.ts";
 import "io/test.ts";
 import "http/server_test.ts";
 import "http/file_server_test.ts";
@@ -23,6 +22,7 @@ import "media_types/test.ts";
 import "multipart/formfile_test.ts";
 import "multipart/multipart_test.ts";
 import "prettier/main_test.ts";
+import "promises/test.ts";
 import "testing/test.ts";
 import "textproto/test.ts";
 import "ws/test.ts";
